EQ1-207 refractor new contrat
